move utils to datastacks
