+++ de_templates/ingest/Ingest_SourceType_SourceName/tests/end_to_end/features/steps/azure_data_ingest_steps.py.jinja
@@ -1,4 +1,4 @@
-from utils.test.e2e.shared_steps import (  # noqa: F401 flake 8
+from datastacks.behave.shared_steps import (  # noqa: F401 flake 8
     trigger_adf_pipeline,
     poll_adf_pipeline,
     pipeline_has_finished_with_state,
