+++ de_workloads/ingest/Ingest_AzureSql_Example/tests/unit/test_azuresql_example.py
@@ -1,6 +1,6 @@
 from jsonschema import validate
 
-from utils.config_utils import (
+from datastacks.config_utils import (
     config_uniqueness_check,
     load_config_as_dict,
     load_configs_as_list,
