--- utils/README.md
@@ -1,3 +1,0 @@
-# Python utilities package
-
-`utils` package contains helper functions used across the solution for local testing.
